% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPTest.R
\name{bp_test}
\alias{bp_test}
\title{BP test for outliers}
\usage{
bp_test(
  data,
  alpha = 0.05,
  distribution = "norm",
  alternative = "two.sided",
  pvalue = FALSE,
  pvalue_gen = 10000
)
}
\arguments{
\item{data}{A given data}

\item{alpha}{The significance level}

\item{distribution}{The distribution name (default: normal distribution)}

\item{alternative}{The alternatives "two.sided"/"greater"/"less" (default: two.sided)}

\item{pvalue}{The indicator to return p-value (TRUE or FALSE (default)).
The p-value is calculated using simulations SO IT MIGHT TAKE SOME TIME}

\item{pvalue_gen}{The number of generations for pvalue calculation; default=10000 takes up to 5 minutes}
}
\value{
The list containing main statistic U, statistics U_i,  ID of outliers, estimates loc and scale,
and number of identified outliers
}
\description{
The calculation of the test statistic U and the p-value of the BP test for
outliers and finding of observations declared by the test as outliers.
}
\examples{
set.seed(12)
x <- rnorm(100)
ks.test(x, "pnorm") # check normality
x[12:20] <- 4
ks.test(x, "pnorm") # check normality, not normal data
bp <- bp_test(x, pvalue = TRUE)
bp
x_after <- x[!bp$outlier]
ks.test(x_after, "pnorm") # check normality, after outliers removal data normal again

set.seed(12)
x <- rcauchy(100)
ks.test(x, "pcauchy") # check cauchy distribution
x[12:22] <- 500
ks.test(x, "pcauchy") # check cauchy, not cauchy data
bp <- bp_test(x, distribution = "cauchy", pvalue = TRUE)
bp
x_after <- x[!bp$outlier]
ks.test(x_after, "pcauchy") # check cauchy, after outliers removal data cauchy again

}
\keyword{BP}
\keyword{test}
