% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\docType{data}
\name{example1}
\alias{example1}
\title{Illustrative example of outliers identification}
\format{Data are a list contains 20 values}
\usage{
example1
}
\description{
The example is considered in the article V. Bagdonavicius, L. Petkevicius
"Multiple outlier detection tests for parametric models"
A dataset contains 20 observations among them 7 are outliers (1st-3rd and 17th-20th).
}
\examples{
bp_test(example1)
}
\keyword{datasets}
