% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example2.R
\docType{data}
\name{example2}
\alias{example2}
\title{Illustrative example of outliers identification}
\format{
Data are a list contains 100 values
}
\usage{
example2
}
\description{
The example is considered in the article V. Bagdonavicius, L. Petkevicius
"Multiple outlier detection tests for parametric models"
A dataset contains 100 observations among them 2 are outliers (3rd and 5th).
}
\examples{
bp_test(example2)
}
\keyword{datasets}
