% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_alpha_n}
\alias{get_alpha_n}
\title{Calculation of significance level alpha_n}
\usage{
get_alpha_n(n, alpha = 0.05)
}
\arguments{
\item{n}{A sample size}

\item{alpha}{A significance level, default value 0.05}
}
\value{
alpha_n
}
\description{
For a given sample size n and the significance level alpha (alpha is the
probability for minimum one observation to fall into the outlier region)
the probability alpha_n for an  individual observation  to fall into the
outlier region is calculated.
}
\examples{
get_alpha_n(1) # returns alpha_n = 0.05 since sample size n = 1 and default value of alpha is 0.05

get_alpha_n(1, 0.1) # returns alpha_n = 0.1 since sample size n = 1
}
\keyword{alpha_n}
