% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_an}
\alias{get_an}
\title{constant a_n}
\usage{
get_an(n, m = 0, s = 1, distribution = "norm",
  alternative = "greater")
}
\arguments{
\item{n}{A sample size}

\item{m}{An estimate of location parameter mu (default value (theoretical) 0)}

\item{s}{An estimate of shape parameter sigma (default value (theoretical) 1)}

\item{distribution}{The distribution name}

\item{alternative}{The choice of alternative}
}
\value{
ret The a_n value
}
\description{
The calculation of the constant a_n nessisary for BP test
}
\examples{
get_an(100, distribution = "norm")


}
\keyword{a_n}
\keyword{constant}
