% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_bn}
\alias{get_bn}
\title{constant b_n}
\usage{
get_bn(n, m = 0, s = 1, distribution = "norm", alternative = "greater")
}
\arguments{
\item{n}{A sample size}

\item{m}{An estimate of location parameter mu (default value (theoretical) 0)}

\item{s}{An estimate of shape parameter sigma (default value (theoretical) 1)}

\item{distribution}{The distribution name}

\item{alternative}{The choice of alternative}
}
\value{
ret The b_n value
}
\description{
The calculation of the constant b_n necessary for BP test
}
\examples{
get_bn(100, distribution = "norm")

}
\keyword{b_n}
\keyword{constant}
