% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_critical}
\alias{get_critical}
\title{Critical values of BP test}
\usage{
get_critical(n, s = 5, alpha = 0.05, distribution = "norm",
  alter = "two.sided")
}
\arguments{
\item{n}{A sample size}

\item{s}{A upper limit (default s=5)}

\item{alpha}{The significance level  (default: 0.05)}

\item{distribution}{The distribution name}

\item{alternative}{The choice of alternative ("two.sided"/"greater"/"less")}
}
\value{
ret The critical value of statistics U
}
\description{
The method with estimated aproximations of critical values.
The unimplemented cases using rough approximation instead.
}
\examples{
get_critical(100)


}
\keyword{BP}
\keyword{critical}
\keyword{values}
