% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_main_bp_statistic}
\alias{get_main_bp_statistic}
\title{Main BP statistic}
\usage{
get_main_bp_statistic(data, distribution = "norm",
  alternative = "two.sided")
}
\arguments{
\item{data}{A given data}

\item{distribution}{The distribution name}

\item{alternative}{The choice of alternative}
}
\value{
ret The value of the statistic U
}
\description{
The calculation of the value of the BP test statistic U.
}
\examples{
x <- rnorm(100)
get_main_bp_statistic(x)


}
\keyword{BP}
\keyword{statistic}
