% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_pvalue}
\alias{get_pvalue}
\title{p-value of BP test}
\usage{
get_pvalue(statistic_val, n, distribution = "norm",
  alternative = "two.sided", gen_num = 10000)
}
\arguments{
\item{statistic_val}{A given BP statistics value}

\item{n}{The size of the sample}

\item{distribution}{The distribution name}

\item{alternative}{The chooise of alternative}

\item{gen_num}{The number of simulations to used in order to calculate p-value.}
}
\value{
ret The p-value of given BP statistics
}
\description{
The use simulation and for given conditions calculate exact p-value.
Since simulation is used it might take few minutes.
}
\examples{
x <- rnorm(100)
bp <- bp_statistic(x)
get_pvalue(bp$Test_statistic_U, 100)

}
\keyword{BP}
\keyword{p-value}
