% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_robust_estimates}
\alias{get_robust_estimates}
\title{Robust estimates for location-scale families of distributions}
\usage{
get_robust_estimates(x, distribution = "norm")
}
\arguments{
\item{x}{A sample}

\item{distribution}{The distribution name}
}
\value{
ret The list structure with location and scale -  the estimates of location and scale parameters.
}
\description{
The robust estimates of the scale (Q_n) and location parameters for
location-scale families of distributions.
Implemented distributions: normal, gumbel, cauchy, laplace, logistic
}
\examples{
x <- rnorm(100)
estimates <- get_robust_estimates(x)
estimates$location
estimates$scale

x <- rcauchy(100)
estimates <- get_robust_estimates(x, "cauchy")
estimates

x <- rlogis(100)
estimates <- get_robust_estimates(x, "logis")
estimates

}
\keyword{estimator}
\keyword{robust}
