% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{investigate_sample}
\alias{investigate_sample}
\title{data investigation for outliers}
\usage{
investigate_sample(data, alpha = 0.05, verbose = FALSE)
}
\arguments{
\item{data}{A given sample}
}
\description{
The procedure to find outliers and then check the goodness-of-fit with
given distribution.
}
\examples{
investigate_sample(example1)

}
