% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\name{get_critical}
\alias{get_critical}
\title{Critical values of BP test}
\usage{
get_critical(n, s = 5, alpha = 0.05, distribution = "norm",
  alternative = "two.sided", gen_num = 10000)
}
\arguments{
\item{n}{A sample size}

\item{s}{A upper limit (default s=5)}

\item{alpha}{The significance level  (default: 0.05)}

\item{distribution}{The distribution name}

\item{alternative}{The choice of alternative ("two.sided"/"greater"/"less")}

\item{gen_num}{The number of simulations to used in order to calculate p-value.}
}
\value{
ret The critical value of statistics U
}
\description{
The method with estimated aproximations of critical values.
For two.sided alternatives the exact approximations were provided.
For less common configurations the simulation is applied and exact critical value is calculated.
}
\examples{
get_critical(100)


}
\keyword{BP}
\keyword{critical}
\keyword{values}
