% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersTests.r
\docType{package}
\name{outliersTests}
\alias{outliersTests}
\alias{outliersTests-package}
\title{outliersTests: A package containing statistical tests of identification
unknown number of outliers}
\description{
Statistical test for various location-scale family distributions
normal, logistic, cauchy, laplace to test does sample contain outliers and
identifying those outliers in sample.
}
\section{outliersTests functions}{



\code{\link{bp_test}}: The calculation of the test statistic U and the p-value of the BP test for
outliers and finding of observations declared by the test as outliers.

\code{\link{get_robust_estimates}}:  The robust estimates of the scale (Q_n) and location parameters for
location-scale families of distributions.
Implemented distributions: normal, gumbel, cauchy, laplace, logistic
}

\section{outliersTests package data}{


\code{\link{example1}}: - example data
}

